requirements.txt
# requirements.txt

# El Core de Django
django

# Conector de Base de Datos
mysqlclient
pipmysql

# Hashing de contraseñas
bcrypt

# Para leer archivos .env 
python-dotenv

# Para la API
djangorestframework
djangorestframework-simplejwt==5.3.1